Add more details Admin UI
